Rewrite copr-builder into an installable python package
