+++ copr_builder/errors.py
@@ -1,0 +1,19 @@
+
+class CoprBuilderError(Exception):
+    pass
+
+
+class GitError(CoprBuilderError):
+    pass
+
+
+class SRPMBuilderError(CoprBuilderError):
+    pass
+
+
+class CoprBuilderAlreadyFailed(CoprBuilderError):
+    pass
+
+
+class CoprBuilderConfigurationError(CoprBuilderError):
+    pass
